Harness v2: PreToolUse type-vocab gate + audit-ratchet
.claude/hooks/pretooluse-type-vocab.sh:
  Reads the Write/Edit payload from stdin. For each `pub struct/enum/
  trait/type NAME` in the proposed content, checks the type-vocabulary
  registry. If NAME is enforce-mode and the target path isn't its
  canonical owner, exits 2 to reject the tool call. The agent gets the
  full remediation message (canonical owner path, fix suggestion) in
  stderr before it spends tokens on the bad work.

Wired into PreToolUse:Write|Edit in settings.json alongside the existing
verify-gate.sh. Both hooks run; either can reject.

harness/check_audit_ratchet.sh + harness/audit-ratchet.lock:
  Counts mode=audit entries in type-vocabulary.tsv; fails if the count
  exceeds the value in the lock file. Lock is 7 (current state). New
  entries must enter as mode=enforce. Existing audit entries can
  graduate to enforce (which lets architect decrement the lock) or stay
  — the count is strictly non-increasing over project lifetime.

Verified with three test fixtures:
  - non-vocab name → allow
  - canonical name in wrong owner → BLOCK with remediation
  - canonical name in canonical owner → allow

The v2 deferred item (generalize check_type_vocabulary.py into a
shared harness/lib/registry.py) is polish, not a blocker for current
work; will land when we add a second registry (signatures, errors).

+++ crates/lua-stdlib/src/table_lib.rs
@@ -91,7 +91,7 @@ fn check_tab(state: &mut LuaState, arg: i32, what: u32) -> Result<(), LuaError> 
     }
 
     if ok {
-        state.pop_n(n);
+        state.pop_n(n as usize);
         Ok(())
     } else {
         // C: luaL_checktype(L, arg, LUA_TTABLE) — forces a type error
